appviews: improve transition state of expandable content w/ backgrounds
for perf and timing reasons, we don't render expandable content in some
cases, which means while the screens are in transition, the video and
animated gif from the prior screen disappeared, which was a little sad.
it would be very complex to render these for a half-second longer, so
this fix basically just puts reasonably matching backgrounds behind the
assets, which makes them not nearly as noticable, and is definitely good
enough. see https://github.com/gertrude-app/project/issues/270